feat: Make 'xtask check' a 'cargo check' wrapper
We already had an 'xtask check' subcommand that checks crate
metadata against some rules we want to enforce around things
like licensing. I've renamed that one to 'xtask validate'
and updated 'xtask ci' (which calls it) to use the new name.

Now 'xtask check' is a wrapper for 'cargo check' in the same
way that 'xtask build' is a wrapper for 'cargo build'.
